Allow inspection of disabled plugins by not stopping resource loading
The differences between a DISABLED plugin and a STARTED plugin are:

1. a STARTED plugin has executed Plugin.start(), a DISABLED plugin has not
2. a STARTED plugin may contribute extension instances, a DISABLED plugin may not

DISABLED plugins still have valid classloaders and their classes can be manually
loaded and explored, but the resource loading - which is important for inspection
has been handicapped by the DISABLED check.

Instead of preventing loading the extension indexes for DISABLED plugins, the
extension finder should only return ExtensionWrappers for STARTED plugins.